pyproject.toml
[project]
name = "video-to-reels"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "altair==4.0.0",
    "anthropic>=0.86.0",
    "ffmpeg-python>=0.2.0",
    "filetype>=1.2.0",
    "huggingface-hub>=1.7.2",
    "imageio-ffmpeg>=0.6.0",
    "librosa>=0.11.0",
    "loguru>=0.7.3",
    "moviepy>=1.0.3",
    "numpy>=2.4.3",
    "openai>=2.29.0",
    "openai-whisper>=20250625",
    "opencv-contrib-python>=4.13.0.92",
    "opencv-python>=4.13.0.92",
    "pandas>=2.3.3",
    "pillow>=12.1.1",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.2",
    "pytube>=15.0.0",
    "requests>=2.32.5",
    "scipy>=1.17.1",
    "standard-imghdr>=3.13.0",
    "streamlit==1.55.0",
    "streamlit-extras>=0.1.5",
    "torchvision>=0.26.0",
    "tqdm>=4.67.3",
    "ultralytics>=8.4.26",
    "whisper>=1.1.10",
    "yt-dlp>=2026.3.17",
    "zai-sdk>=0.2.2",
]
